requirements.txt
torch
torchvision
matplotlib
tqdm
argparse
Pillow
